run_qmake2cmake: Improve handling of multiple .pro files per directory
If there are multiple .pro files per directory, run_qmake2cmake now
distinguishes the following situations:

1. There's only one SUBDIRS file.  Then, only the SUBDIRS .pro file is
converted.  It's assumed that the other .pro files are added to the
SUBDIRS variable, and they are skipped.

2. There is no or multiple SUBDIRS .pro file(s).  Only the first .pro
file is being converted.  The user must run qmake2cmake on each single
file with the -o argument and integrate the result into the build
system as needed.

Change-Id: If6b8c13baa98262804580715fed8cc50ccb7b851
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>