pro2cmake: Set a project version for examples
For example projects, change the generated project() call to
   project(foo VERSION 1.0 LANGUAGES CXX)

Some CMake API derives default values from the project version, and it's
generally advisable to set a project version number.

The version number is read from qmake's VERSION variable.  That's
actually a target version number, but it might be the right thing in
most cases.  Fall back to version 1.0 if VERSION is not set.

Task-number: QTBUG-96799
Change-Id: Ia0c551cf62621eb217e1dd541dcbd8945f78138e
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>