pro2cmake: Re-work how SUBDIRS are handled for user projects
For user projects, we create a "top-level prelude" only in the top-level
CMakeLists.txt.  This prelude consists of the project() call, several
initializations and the find_package calls. CMakeLists.txt files further
down in the SUBDIRS hierarchy don't get this prelude.

As a consequence, only the top-level CMakeLists.txt contains
find_package calls for all packages that are used in the whole project.

Task-number: QTBUG-96799
Change-Id: I7d03985f2fdc943f95294fca85075fcdb875bdf8
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>