run_qmake2cmake: Convert the main project file first
We now convert the top-level project file first.  This ensures that the
SUBDIRS analysis takes place before sub-projects are converted.  The
remaining files are converting in parallel as before.

Implementation note: the all_files list is sorted by project directory
in such a way that the top-level project file comes first.

Change-Id: Ic95e5ca2d1e5e204d65279c573493bf3a36d2318
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>