Add Coin CI instructions for Linux and macOS
The instructions create a python virtual environment,
install dependencies and run make test which runs
mypy, flake8, black and pytest on Linux and macOS
hosts.

The Windows instructions also work, except for the fact that make
fails to run on Windows because the Makefile contains
globs like *.py inside, which jom doesn't like.

That will be fixed later.

Change-Id: I930dacb979978e4781cb247e8d1adb4d4affcea9
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Toni Saario <toni.saario@qt.io>