CI: Repair pip call on Windows
The call to pip.exe was broken on Windows:
Error exec: "pip.exe": cannot run executable found relative to current directory

And this is, because of two things.

1. The go version that's used to build coin was updated to 1.20, and
1.19 introduced a security measure that prohibits executing programs
relative to the current directory.

2. Our build dir on Windows doesn't have a drive letter and is
apparently considered a relative path by golang.

We're adding the drive-letter-less path into the PATH environment
variable and call exec('pip.exe') from go. This results in the error
above.

Fix this by ensuring that the build directory added to PATH has a drive
letter. This is done by using the "agent variable" flavor of the build
directory.

Change-Id: Iefa6d4fd7ada6e64ef74657314eb0659842d7711
Reviewed-by: Toni Saario <toni.saario@qt.io>